+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/createaccountingjournalentry.go
@@ -50,8 +50,6 @@ type CreateAccountingJournalEntryResponse struct {
 	AccountingCreateJournalEntryResponse *shared.AccountingCreateJournalEntryResponse
 	// HTTP response content type for this operation
 	ContentType string
-	// The request made is not valid.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -70,13 +68,6 @@ func (o *CreateAccountingJournalEntryResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *CreateAccountingJournalEntryResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *CreateAccountingJournalEntryResponse) GetStatusCode() int {