+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/getcommercepayment.go
@@ -42,8 +42,6 @@ type GetCommercePaymentResponse struct {
 	CommercePayment *shared.CommercePayment
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -62,13 +60,6 @@ func (o *GetCommercePaymentResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *GetCommercePaymentResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *GetCommercePaymentResponse) GetStatusCode() int {