+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/listcommercepayments.go
@@ -81,8 +81,6 @@ type ListCommercePaymentsResponse struct {
 	CommercePayments *shared.CommercePayments
 	// HTTP response content type for this operation
 	ContentType string
-	// Your `query` parameter was not correctly formed
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -101,13 +99,6 @@ func (o *ListCommercePaymentsResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *ListCommercePaymentsResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *ListCommercePaymentsResponse) GetStatusCode() int {